Switch to golang container for building
I am getting `go get: no install location for directory
/home/runner/work/make-the-journey-map/make-the-journey-map outside
GOPATH`. Looking at the raised issues, it seems I would need to set up
the Go environment before I can get dependencies and build.

I do not have the experience to set it up, so I am trying to switch to
the official golang Docker image, which is supposed to be set up to be
ready to use.