Introduce dummy graph output to the main command
To be used as `go run . -graph`, it will print a textual representation
of a directed graph.

The first implementation will probably be a simplistic Graphviz or
Cypher output.

Note: To preserve the ability to exit with a non-zero status without
traversing through the nodes, I am now returning `(bool, JourneyMap)`
from `LoadMap()`.