Fallback to a null-object Stage for dead links
If the map has validation errors, I still attempt to render the graph.

Currently, if the errors are related to dead/malformed links,
`link.stage` will be `nil` and panics occur when I try to access those.

Instead, I will use a dummy `Stage` with "error" values. The program
will render the graph without panics but still exit non-zero. This
partial graph can still be useful for users (it was useful for me during
editing).