Require "Culture of Automation" for delivery
I feel the "relates to" connection does not give any value at this
point. Consider:

"To do continous delivery, you need to have a culture of automation"

and

"The culture of automation relates to the discipline of continous
delivery"

The latter sentence is vague, so I changed the relation to represent the
former.