Include generated graph in the readme
When I was viewing the last commit in GitHub, I wished there was a
visual comparison between the before and after states.

This commit is the first step towards that: it sets up git LFS for
storing PNG images and embeds the current view in `docs/map.png`.

I already have a `.git/hooks/pre-commit` script that always generates
this image for me. The next step will be including something similar.