Reduce inconsistency in graph output
I got different graphs with each `go run . -graph` command. I needed
consistency.

The `filepath.Walk` traverses files in lexical order, but the output
gets randomised due to the map of ids to `*Stage`s.

This change switches the map of stages to an array of stages to maintain
ordering.

The trade-off is in the `resolveStage` function which now has to loop
through the array to lookup stages by ids. Currently, this is manageable
due to the size of the map. Later, it has to be reviewed.